+++ casic/__init__.py
@@ -1,4 +1,4 @@
 """CASIC - CAN Stack Integrity Checker."""
 
 __all__ = ["__version__"]
-__version__ = "0.0.3"
+__version__ = "0.0.5"